Map conversations to channels, not just to users
This way, if we get a message event for a message the client user sent (where
the message's user is *us* and so doesn't definitively link it to a
conversation), we can still add it to the right conversation.

Also, this is more general, since not all conversations will necessarily go
with a single specific friend - group DM's for instance. For now, we aren't
handling the group DM case but might later.